requirements.txt
python-dotenv==1.0.0
line-bot-sdk>=3.10.0
schedule==1.2.1
pillow==11.2.1
requests>=2.32.3
python-dateutil==2.8.2
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver_manager==4.0.1
dnspython==2.6.0
google-generativeai>=0.3.1
flask==2.3.3
gunicorn==21.2.0
whitenoise==6.6.0
